Update dockerfile, add requirements.txt and change dependabot
Update dockerfile and add requirements.txt:
This allows for easier management and allows for dependabot to auto-open a PR when the pip packages are updated lowering management time.

Have also added the option for chmod of the entry.sh to be done during the copy stage rather than later, allowing this step to be cached when required, lowering build time and layer size.