Remove unnecessary VOLUME definitions in Dockerfile
This removes VOLUME definitions for locations in the container where we
expect the user to bind their own storage locations or locations where
we don't expect the build process to place files.
The impact of this change is that users using docker compose will not
see unused but declared VOLUMEs being created as anonymous volumes.

Signed-off-by: Felix Kaechele <felix@kaechele.ca>